@@ -1,23 +1,17 @@
-plugins {
-    id 'com.android.application'
-    id 'kotlin-android'
-}
-
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'com.google.gms.google-services'
 
 android {
-    compileSdkVersion 30
+    compileSdk 33
 
     defaultConfig {
         applicationId "com.ados.mstrotrematch2"
-        minSdkVersion 21
-        targetSdkVersion 30
-        multiDexEnabled true
-        versionCode 19
-        versionName "1.19"
+        minSdk 21
+        targetSdk 33
+        versionCode 32
+        versionName "1.2.2"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
