@@ -1,17 +1,17 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
-    ext.adfit_version = '3.0.12'
-    ext.play_service_version = '16.0.0'
-    ext.kotlin_version = '1.3.61'
+    ext.adfit_version = '3.13.6'
+    ext.play_service_version = '17.0.0'
+    ext.kotlin_version = '1.6.20'
     ext.anko_version='0.10.5'
     repositories {
         google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.1.1'
+        classpath 'com.android.tools.build:gradle:7.3.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'com.google.gms:google-services:4.3.4'
+        classpath 'com.google.gms:google-services:4.3.14'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -21,7 +21,7 @@ allprojects {
     repositories {
         google()
         jcenter()
-        maven { url 'http://devrepo.kakao.com:8088/nexus/content/groups/public/' }
+        maven { url 'https://devrepo.kakao.com/nexus/content/groups/public/' }
         maven {
             url "s3://repo.cauly.net/releases"
             credentials(AwsCredentials) {
