@@ -3,21 +3,27 @@ package com.ados.mstrotrematch2.dialog
 
 import android.app.Dialog
 import android.content.Context
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
 import android.os.CountDownTimer
 import android.view.View
 import com.ados.mstrotrematch2.R
+import com.ados.mstrotrematch2.databinding.EventDialogBinding
 import com.ados.mstrotrematch2.model.EventDTO
 import com.bumptech.glide.Glide
-import kotlinx.android.synthetic.main.event_dialog.*
 
 class EventDialog(context: Context, var event : EventDTO) : Dialog(context), View.OnClickListener {
 
-    private val layout = R.layout.event_dialog
+    lateinit var binding: EventDialogBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(layout)
+        binding = EventDialogBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        window!!.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
+
         init()
 
         var limitTime = event.limit?.time
@@ -34,22 +40,22 @@ class EventDialog(context: Context, var event : EventDTO) : Dialog(context), Vie
                 var min = (totalsec % 3600) / 60
                 var sec = totalsec % 60
 
-                text_timer.text = "[${String.format("%02d",hour)}시${String.format("%02d",min)}분${String.format("%02d",sec)}초] 후 티켓이 사라집니다."
+                binding.textTimer.text = "[${String.format("%02d",hour)}시${String.format("%02d",min)}분${String.format("%02d",sec)}초] 후 티켓이 사라집니다."
             }
 
         }.start()
     }
 
     private fun init() {
-        button_ok.setOnClickListener(this)
+        binding.buttonOk.setOnClickListener(this)
 
-        text_msg.text = event.title?.replace("\\n","\n")
+        binding.textMsg.text = event.title?.replace("\\n","\n")
 
-        Glide.with(img_ticket.context)
+        Glide.with(binding.imgTicket.context)
             .asBitmap()
             .load(R.drawable.ticekt) ///feed in path of the image
-            .fitCenter()
-            .into(img_ticket)
+            .optionalFitCenter()
+            .into(binding.imgTicket)
     }
 
     override fun onClick(v: View) {
