@@ -6,20 +6,30 @@ import android.content.Context
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
+import android.view.ViewGroup
 import com.ados.mstrotrematch2.R
+import com.ados.mstrotrematch2.databinding.LoadingDialogBinding
 import com.bumptech.glide.Glide
-import kotlinx.android.synthetic.main.loading_dialog.*
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 
 class LoadingDialog(context: Context) : Dialog(context) {
 
-    private val layout = R.layout.loading_dialog
+    lateinit var binding: LoadingDialogBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(layout)
+        binding = LoadingDialogBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
+        window?.setLayout(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT)
         window!!.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
 
-        Glide.with(context).load(R.raw.loading).into(img_loading)
+        Glide.with(context).load(R.raw.loading)
+            .transform(RoundedCorners(30))
+            .into(binding.imgLoading)
+    }
+
+    override fun onBackPressed() {
+        //super.onBackPressed()
     }
 }
